Create TestGameMenu
You don't need the "package Project;"

There are still some adjustments I need to make but this is what I have so far. I plan to resize the gameMenuPanel and gameSettingsDialog so that they display things properly. Additionally, I plan to try to add a chess piece icon onto the gameMenuPanel and may get rid of the rulesButton instead replacing it with a scrollPanel that lists the rules in the open space to the right.

I'm not sure how you plan to implement the game logic in order to move the pieces around the board or if I need to desing the gameBoardPanel a certain way (i.e. with multiple JPanels, or something else) so just let me know.

I do have a folder full of chess board assets including light pieces, dark pieces, dark board tiles, and light board tiles. But as mentioned above, not sure how I need to add them in order to make them usable for you.